requirements.txt
azure-ai-documentintelligence
openai
python-dotenv
cohere
chromadb>=0.5.0 # Compatible with Python 3.11.9, avoids known issues with latest versions
pydantic
pydantic-settings
fastapi
tqdm
uvicorn[standard]
numpy==2.0.0
langchain
langchain-community